refactor: change naming convention for pinia store defines & file names
